# Fix rounding in optimize_ticks