Undoes part of #83, to correctly compute the number of significant digits.
Before:
```
plot(15:24,1:10)
xlims!(14,25)
```
![tmp](https://user-images.githubusercontent.com/4170948/80613457-fe2d9c80-8a45-11ea-9999-3185eb3c13f2.png)
After:
![tmp](https://user-images.githubusercontent.com/4170948/80613553-1bfb0180-8a46-11ea-8440-2010ffc60e32.png)

This should fix #86, though I can't verify that right now, because when I run `Plots` tests locally it fails due to missing `ImageIO`.