@asinghvi17 2020-04-29
Before merging, we should make sure that there is no significant slowdown for interactive use...could you benchmark this a little with some test data?