@daschw 2020-04-29
Awesome, I can confirm that this fixes #86 
If you want to add it to the tests, here are the numbers:
```julia
julia> PlotUtils.optimize_ticks(-1.0, 13.3166247903554, k_min = 4, k_max = 8)
([0.0, 2.0, 4.0, 6.0, 8.0, 10.0, 12.0], -1.0, 13.3166247903554)
```
Before this it returned 0:2:10