@yha 2020-04-29
> Before merging, we should make sure that there is no significant slowdown for interactive use...could you benchmark this a little with some test data?

I found no significant change with this benchmark:
```
using BenchmarkTools
@benchmark PlotUtils.optimize_ticks(x,y) setup=((x,y)=minmax(rand(-100:100,2)...))
```

Before:
```
BenchmarkTools.Trial: 
  memory estimate:  72.11 KiB
  allocs estimate:  1481
  --------------
  minimum time:     381.185 μs (0.93% GC)
  median time:      482.575 μs (0.85% GC)
  mean time:        465.402 μs (1.10% GC)
  maximum time:     530.127 μs (1.38% GC)
  --------------
  samples:          16
  evals/sample:     685
```
After:
```
BenchmarkTools.Trial: 
  memory estimate:  71.45 KiB
  allocs estimate:  1467
  --------------
  minimum time:     382.179 μs (1.03% GC)
  median time:      440.826 μs (0.93% GC)
  mean time:        453.471 μs (1.19% GC)
  maximum time:     535.145 μs (1.48% GC)
  --------------
  samples:          17
  evals/sample:     676
```

> If you want to add it to the tests, here are the numbers:

Added to the tests, (with the rounded `(x,y)=(-1.0,13.0)`, which also failed).

> Before this it returned 0:2:10

Actually, it returned [0:2:10;10], with the 10 appearing twice (the second time is from rounding of 12 to 1 significant digit).